requirements.txt
streamlit
PyMuPDF
python-docx
easyocr
pillow
numpy